+++ CHANGELOG.md
@@ -87,6 +87,6 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 [TypeScript]: http://www.typescriptlang.org
 [TypeScript declaration file]: https://www.typescriptlang.org/docs/handbook/writing-declaration-files.html
 
-[2.0.1]: https://github.com/pelotoncycle/frame-throttle/compare/v1.1.0...v2.0.1
-[2.0.0]: https://github.com/pelotoncycle/frame-throttle/compare/v1.1.0...v2.0.0
 [1.1.0]: https://github.com/pelotoncycle/frame-throttle/compare/v1.0.0...v1.1.0
+[2.0.0]: https://github.com/pelotoncycle/frame-throttle/compare/v1.1.0...v2.0.0
+[2.0.1]: https://github.com/pelotoncycle/frame-throttle/compare/v1.1.0...v2.0.1